Add gauge level filtering

+++ src/types/pagination.rs
@@ -4,14 +4,14 @@ use serde::{Deserialize, Serialize};
 #[serde(rename_all = "camelCase")]
 pub struct PaginationQuery {
     #[validate(range(min = 1, max = 10000))]
-    pub page: usize,
+    pub page: i64,
     #[validate(range(min = 1, max = 10))]
-    pub limit: usize,
+    pub limit: i64,
 }
 
 #[derive(Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Pagination {
-    pub page: usize,
-    pub next_page: Option<usize>,
+    pub page: i64,
+    pub next_page: Option<i64>,
 }
